refactor: needn't FnMut for next clousre
when an operator next callback is a FnMut means the evniroment vaiable has been mut borrow in all
stream lifetime. In this period, stream can modify the enviroment variable, but the customer can't
read it. It's not so useful, but limit rx_rs can't hold two reference of the callback.