Do not subscribe anew on each new group
The existing implementation of the `group_by` functionality contained a
bug where on each new group, a new `subscribe` chain was initiated,
possibly up to the value source (see #201).
This commit fixes the problem by emitting one `subject` per group, which
acts as both `Observer` for the upstream processing pipeline, and as
`Observable` for the downstream pipeline.
Implementing `group_by` in this way means that each value travels down
the stream, being rerouted by `GroupByObserver` to the correct
`subject`.